pyproject.toml
[project]
name = "testudot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4",
    "fastapi",
    "httpx",
    "pydantic<2.0",
    "python-dotenv",
    "upstash-redis",
    "rich",
    "typer",
    "schedule",
    "uvicorn",
]

[dependency-groups]
dev = [
    "workers-py",
    "workers-runtime-sdk",
    "ruff",
]
